requirements.txt
Django==4.2.16
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple==3.8
django-simple-history==3.7.0

